+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Upgrade the `org.realityforge.javax.annotation` artifact to version `1.0.1`.
+* Move some service logic to being default methods on the `TimeService` interface. This makes it easier to re-implement the service logic within tests.
 
 ### [v0.01](https://github.com/realityforge/timeservice/tree/v0.01) (2019-04-24) · [Full Changelog](https://github.com/realityforge/timeservice/compare/c97326165a2eaacc2610d8a11c03a9ba01f24ecd...v0.01)